requirements.txt
flask
pymysql
python-dotenv
pandas
openpyxl
xlsxwriter
sqlalchemy
pytest
pytest-cov
pylint
mock
numpy